pyproject.toml
[project]
name = "ethan-llm-proxy-gateway"
version = "0.1.0"
description = "LLM proxy gateway with routing, fallbacks, redaction, and audit events."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4",
  "mypy>=1.8",
  "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
